refactor: move parser to internals
