+++ src/main/scala/no/sysco/middleware/kafka/event/collector/cluster/ClusterManager.scala
@@ -6,7 +6,8 @@ import akka.actor.{Actor, ActorLogging, ActorRef, Props}
 import io.opencensus.scala.Stats
 import io.opencensus.scala.stats.Measurement
 import no.sysco.middleware.kafka.event.collector.cluster.NodeManager.ListNodes
-import no.sysco.middleware.kafka.event.collector.model.{Cluster, ClusterDescribed, NodesDescribed, Parser}
+import no.sysco.middleware.kafka.event.collector.internal.Parser
+import no.sysco.middleware.kafka.event.collector.model.{Cluster, ClusterDescribed, NodesDescribed}
 import no.sysco.middleware.kafka.event.proto.collector._
 
 import scala.concurrent.ExecutionContext
