+++ src/main/scala/no/sysco/middleware/kafka/event/collector/cluster/NodeManager.scala
@@ -3,6 +3,7 @@ package no.sysco.middleware.kafka.event.collector.cluster
 import akka.actor.{Actor, ActorLogging, ActorRef, Props}
 import io.opencensus.scala.Stats
 import io.opencensus.scala.stats.Measurement
+import no.sysco.middleware.kafka.event.collector.internal.Parser
 import no.sysco.middleware.kafka.event.collector.model._
 import no.sysco.middleware.kafka.event.proto.collector.{NodeCreated, NodeEvent, NodeUpdated}
