+++ src/main/scala/no/sysco/middleware/kafka/event/collector/topic/TopicManager.scala
@@ -6,6 +6,7 @@ import akka.actor.{Actor, ActorLogging, ActorRef, Props}
 import akka.stream.ActorMaterializer
 import io.opencensus.scala.Stats
 import io.opencensus.scala.stats.Measurement
+import no.sysco.middleware.kafka.event.collector.internal.Parser
 import no.sysco.middleware.kafka.event.collector.model._
 import no.sysco.middleware.kafka.event.proto.collector.{TopicCreated, TopicDeleted, TopicEvent}
